fix: fix integer precision loss from casting long (int64) to int (int32)

+++ lib/src/main/java/dev/doglog/DogLog.java
@@ -120,8 +120,8 @@ public class DogLog {
     logExtrasIfPeriodElapsed();
   }
 
-  /** Log an int. */
-  public static void log(String key, int value) {
+  /** Log a long. */
+  public static void log(String key, long value) {
     if (enabled) {
       logger.queueLog(key, value);
     }
